requirements.txt
pandas
biopython
rdkit
tqdm
scikit-learn
requests
jupyter
matplotlib
seaborn
